Pequenas correções

+++ exercicio1.py
@@ -8,12 +8,13 @@ def tresNumeros(num1, num2, num3):
         return True
     else:
         return False
+
 try:
-    num1 = int(input("Informe um número inteiro qualquer: "))
+    num1 = float(input("Informe um número inteiro qualquer: "))
 
-    num2 = int(input("Informe um número inteiro qualquer: "))
+    num2 = float(input("Informe um número inteiro qualquer: "))
 
-    num3 = int(input("Informe um número inteiro qualquer: "))
+    num3 = float(input("Informe um número inteiro qualquer: "))
 
     if tresNumeros(num1, num2, num3) == True:
         print(f"A soma entre {num1} e {num2} é igual à {num3}")
